Revert "Use OTel library from tracing while keeping Jaeger exporter config (#…"
This reverts commit 699a122aba83da75ccef884d63af468da4f5127e.

+++ pkg/streamingpromql/query.go
@@ -590,7 +590,7 @@ func (q *Query) Exec(ctx context.Context) *promql.Result {
 
 		msg = append(msg,
 			"msg", "query stats",
-			"estimatedPeakMemoryConsumption", int64(q.memoryConsumptionTracker.PeakEstimatedMemoryConsumptionBytes()),
+			"estimatedPeakMemoryConsumption", q.memoryConsumptionTracker.PeakEstimatedMemoryConsumptionBytes(),
 			"expr", q.originalExpression,
 		)
